+++ src/microsim/MSLink.cpp
@@ -740,7 +740,7 @@ MSLink::blockedByFoe(const SUMOVehicle* veh, const ApproachingVehicleInformation
         gDebugFlag6 = ((ego == nullptr || ego->isSelected()) && (veh == nullptr || veh->isSelected()));
 #endif
         const SUMOTime fatb = computeFoeArrivalTimeBraking(arrivalTime, veh, avi.arrivalTime, impatience, avi.dist, foeArrivalSpeedBraking);
-        foeArrivalTime = (SUMOTime)((1.0 - impatience) * avi.arrivalTime + impatience * fatb);
+        foeArrivalTime = (SUMOTime)((1. - impatience) * (double)avi.arrivalTime + impatience * (double)fatb);
 #ifdef MSLink_DEBUG_OPENED
         if (gDebugFlag6) {
             std::cout << SIMTIME << " link=" << getDescription() << " ego=" << ego->getID() << " foe=" << veh->getID()
